Simplify config selection
This commit changes how configs are found and read in the following way:

First, we no longer look in the `$XDG_DATA_DIRS`.  Any files here should
be considered *templates*, and not actual configs to be used.

Second, we only read *one* config file.  This means that it's no longer
possible to e.g. put a file of defaults in a host-wide
`$XDG_CONFIG_DIRS`, and override settings from this in a user-specific
config in `$XDG_CONFIG_HOME`.  The original functionality mostly added
complexity to something that should be simple.  Will probably
re-consider the config format as well, but we'll see.